refactor: extract microcompact to astra-turn-core (Phase 22)
Move microcompact.rs (~1,582 LOC) from runtime to astra-turn-core.
Contains message compaction, tool result deduplication, and state-aware pinning.

Import fixes: super::cloud::session_facts → crate::cloud_session_facts

Turn-core: 1754 tests passing
Runtime: 3029 tests + 5 pre-existing remote skill failures
